python/requirements.txt
flask
xgboost
pandas
mysql-connector-python
gunicorn
python-dotenv
